refactor(test): fold merge conformance into MultiIndexMergeEngineTestBase
The previous commit added a standalone MergeConformanceTestBase plus one
Test<FW>MergeConformance subclass per framework. Since every framework
merge-engine test already subclasses MultiIndexMergeEngineTestBase with
the identical merge_engine_class / framework_type / get_connection hooks,
the conformance check belongs there.

Move the parametrized test_merge_conformance (plus the _norm/_multiset
helpers) onto MultiIndexMergeEngineTestBase and drop the standalone base
and all seven conformance subclasses. Non-lazy PolarsMergeEngine had no
multi-index subclass, so give it TestPolarsMergeEngineMultiIndex, which
also grants it the multi-index tests it previously lacked (safe: the lazy
engine subclasses it and already passes them, including union).

Same 56 conformance cases, canonical scenarios unchanged.
