+++ tests/test_plugins/compute_framework/base_implementations/sqlite/test_sqlite_merge_engine.py
@@ -11,9 +11,6 @@ from mloda.provider import BaseMergeEngine
 from mloda_plugins.compute_framework.base_implementations.sqlite.sqlite_framework import _regexp
 from mloda_plugins.compute_framework.base_implementations.sqlite.sqlite_merge_engine import SqliteMergeEngine
 from mloda_plugins.compute_framework.base_implementations.sqlite.sqlite_relation import SqliteRelation
-from tests.test_plugins.compute_framework.test_tooling.merge_conformance.merge_conformance_test_base import (
-    MergeConformanceTestBase,
-)
 from tests.test_plugins.compute_framework.test_tooling.merge_link import make_merge_link
 from tests.test_plugins.compute_framework.test_tooling.multi_index.multi_index_test_base import (
     MultiIndexMergeEngineTestBase,
@@ -355,24 +352,6 @@ def test_execute_sql_raises_value_error_when_no_connection() -> None:
 
 
 class TestSqliteMergeEngineMultiIndex(MultiIndexMergeEngineTestBase):
-    @classmethod
-    def merge_engine_class(cls) -> type[BaseMergeEngine]:
-        return SqliteMergeEngine
-
-    @classmethod
-    def framework_type(cls) -> type[Any]:
-        return SqliteRelation
-
-    def get_connection(self) -> Optional[Any]:
-        if not hasattr(self, "_connection"):
-            self._connection = sqlite3.connect(":memory:")
-            self._connection.create_function("REGEXP", 2, _regexp)
-        return self._connection
-
-
-class TestSqliteMergeConformance(MergeConformanceTestBase):
-    """Cross-framework merge conformance for SqliteMergeEngine."""
-
     @classmethod
     def merge_engine_class(cls) -> type[BaseMergeEngine]:
         return SqliteMergeEngine
